+++ frontend/src/pages/DashboardPageV2/DashboardContainer/VariablesBar/__tests__/resolveVariableSelection.test.ts
@@ -6,7 +6,7 @@ import {
 	configuredDefaultValue,
 	reconcileWithOptions,
 	resolveDefaultSelection,
-} from '../resolveVariableSelection';
+} from '../utils/resolveVariableSelection';
 
 function model(overrides: Partial<VariableFormModel>): VariableFormModel {
 	return { ...emptyVariableFormModel(), ...overrides };
