+++ frontend/src/pages/DashboardPageV2/DashboardContainer/VariablesBar/components/AddVariable/AddVariableFull.tsx
@@ -1,8 +1,8 @@
 import { Plus } from '@signozhq/icons';
 import { Button } from '@signozhq/ui/button';
 
-import { useDashboardStore } from '../store/useDashboardStore';
-import styles from './VariablesBar.module.scss';
+import { useDashboardStore } from '../../../store/useDashboardStore';
+import styles from './AddVariable.module.scss';
 
 /**
  * Full-width labelled "Add variable" button shown in the empty state, before any
