+++ frontend/src/pages/DashboardPageV2/DashboardContainer/store/slices/__tests__/variableFetchSlice.test.ts
@@ -5,7 +5,7 @@ import {
 import {
 	deriveFetchContext,
 	type VariableFetchContext,
-} from '../../../VariablesBar/variableDependencies';
+} from '../../../VariablesBar/utils/variableDependencies';
 import { useDashboardStore } from '../../useDashboardStore';
 
 function model(overrides: Partial<VariableFormModel>): VariableFormModel {
